--- a/Cabal/tests/ParserTests/regressions/indentation.cabal
+++ b/Cabal/tests/ParserTests/regressions/indentation.cabal
@@ -1,0 +1,18 @@
+cabal-version: 3.0
+name: indentation
+version: 0
+description:
+
+  * foo
+
+      * foo-bar
+
+      * foo-baz
+
+  .
+  .
+  .
+  some dots
+
+library
+  default-language: Haskell2010