@phadej 2019-03-28
This has an interesting consequence of making parsing 3x (three times) faster.

```
master % time .../hackage-tests parsec   
Reading index from: /home/ogre/.cabal/packages/hackage.haskell.org/01-index.tar
123149 files processed
 parsec  346,10s user 5,37s system 97% cpu 5:59,84 total

freeTextField % time .../hackage-tests parsec
Reading index from: /home/ogre/.cabal/packages/hackage.haskell.org/01-index.tar
123149 files processed
 parsec  101,18s user 0,75s system 100% cpu 1:41,86 total
```

Please, verify I haven't broken anything. At least `roundtrip` for whole hackage still works, so at least I broke parsing and pretty-printing in a consistent way,.